Maj from test to main

+++ Bank/Bank.hpp
@@ -18,7 +18,7 @@ public:
     static void freeInstance();
     int deposit(size_t idPlayer,int amount);
     void withdraw(size_t idPlayer,int amount);
-    int getBalance(int id){ return wallets[id]->getBalance(); }
-    void setBalance(int id, int num) { wallets[id]->setBalance(num);}
+    int getBalance(size_t id){ return wallets[id]->getBalance(); }
+    void setBalance(size_t id, int num) { wallets[id]->setBalance(num);}
     void playerPaysPlayer(size_t idPlayer1,size_t idPlayer2,int amount);
 };
